[Relay][Text Format] Pretty Printer Smart Inlining (#2881)
